More lint fixes
