+++ client/src/StaticContext.js
@@ -1,0 +1,17 @@
+import { createContext, useContext } from 'react';
+
+export const staticContext = createContext();
+
+export const defaultValue = {
+  authContext: {
+    user: null,
+  },
+  env: {
+    VITE_SITE_TITLE: import.meta.env.VITE_SITE_TITLE,
+    VITE_FEATURE_REGISTRATION: import.meta.env.VITE_FEATURE_REGISTRATION,
+  },
+};
+
+export function useStaticContext() {
+  return useContext(staticContext);
+}
